Tests update

+++ ngraph/test/backend/transpose.in.cpp
@@ -166,18 +166,18 @@ NGRAPH_TEST(${BACKEND_NAME}, transpose_int_data_axes_constant)
     const auto axes_shape = Shape{4};
     const auto output_shape = Shape{3, 4, 2, 1};
 
-    auto data_param = make_shared<op::Parameter>(element::i16, data_shape);
-    auto axes_const = op::Constant::create(element::i32, axes_shape, {2, 3, 0, 1});
+    auto data_param = make_shared<op::Parameter>(element::i32, data_shape);
+    auto axes_const = op::Constant::create(element::i64, axes_shape, {2, 3, 0, 1});
     auto transpose = make_shared<op::Transpose>(data_param, axes_const);
     auto function = make_shared<ngraph::Function>(NodeVector{transpose}, ParameterVector{data_param});
 
-    std::vector<int16_t> data(shape_size(data_shape));
+    std::vector<int32_t> data(shape_size(data_shape));
     std::iota(data.begin(), data.end(), 1);
-    std::vector<int16_t> expected_result{ 1, 13,  2, 14,  3, 15,  4, 16,  5, 17,  6, 18,  7, 19,  8, 20,  9,
+    std::vector<int32_t> expected_result{ 1, 13,  2, 14,  3, 15,  4, 16,  5, 17,  6, 18,  7, 19,  8, 20,  9,
        21, 10, 22, 11, 23, 12, 24};
 
     auto test_case = test::TestCase<TestEngine>(function);
-    test_case.add_input<int16_t>(data_shape, data);
-    test_case.add_expected_output<int16_t>(output_shape, expected_result);
+    test_case.add_input<int32_t>(data_shape, data);
+    test_case.add_expected_output<int32_t>(output_shape, expected_result);
     test_case.run();
 }
